fix unit test
Move method to test_local.py
